# openapi/openapi.yaml
openapi: 3.0.0
info:
  title: Notification Service
  version: 0.0.0
tags:
  - name: Worker
  - name: Messages
paths:
  /messages:
    get:
      operationId: Messages_sent
      parameters:
        - name: limit
          in: query
          required: true
          schema:
            type: integer
          explode: false
        - name: offset
          in: query
          required: true
          schema:
            type: integer
          explode: false
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Message'
        default:
          description: An unexpected error response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      tags:
        - Messages
    post:
      operationId: Messages_add
      parameters: []
      responses:
        '201':
          description: The request has succeeded and a new resource has been created as a result.
        default:
          description: An unexpected error response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      tags:
        - Messages
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Message'
  /worker/start:
    get:
      operationId: Worker_start
      summary: Starts the polling worker
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkerResponse'
        default:
          description: An unexpected error response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      tags:
        - Worker
  /worker/stop:
    get:
      operationId: Worker_stop
      summary: Stops the polling worker
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkerResponse'
        default:
          description: An unexpected error response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      tags:
        - Worker
components:
  schemas:
    Error:
      type: object
      required:
        - message
      properties:
        message:
          type: string
    Message:
      type: object
      required:
        - id
        - to
        - content
        - status
        - createdAt
        - updatedAt
      properties:
        id:
          type: integer
          readOnly: true
        to:
          type: string
          title: Phone number of the recipient
          example: '+40700000000'
        content:
          type: string
          title: Message content
          example: How are you on this beautiful day?
        status:
          type: string
          readOnly: true
        createdAt:
          type: string
          format: date-time
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          readOnly: true
    WorkerResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
servers:
  - url: /api/v1
    description: API v1
    variables: {}